# openapi.json
{
  "components": {
    "schemas": {
...
      "DriftStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "adwin_change_detected": {
            "default": false,
            "title": "Adwin Change Detected",
            "type": "boolean"
          },
          "feature_drift_count": {
            "default": 0,
            "title": "Feature Drift Count",
            "type": "integer"
          },
          "feature_results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Feature Results",
            "type": "array"
          },
          "latest_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Score"
          },
          "score_stream_count": {
            "title": "Score Stream Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "status",
          "score_stream_count"
        ],
        "title": "DriftStatusResponse",
        "type": "object"
      },
...
      "FeatureDriftRequest": {
        "additionalProperties": false,
        "properties": {
          "alpha": {
            "default": 0.01,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Alpha",
            "type": "number"
          },
          "current_rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 100000,
            "minItems": 2,
            "title": "Current Rows",
            "type": "array"
          },
          "reference_rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 100000,
            "minItems": 2,
            "title": "Reference Rows",
            "type": "array"
          }
        },
        "required": [
          "reference_rows",
          "current_rows"
        ],
        "title": "FeatureDriftRequest",
        "type": "object"
      },
      "FeatureDriftResponse": {
        "additionalProperties": false,
        "properties": {
          "adwin_change_detected": {
            "default": false,
            "title": "Adwin Change Detected",
            "type": "boolean"
          },
          "evaluated_features": {
            "title": "Evaluated Features",
            "type": "integer"
          },
          "feature_drift_count": {
            "default": 0,
            "title": "Feature Drift Count",
            "type": "integer"
          },
          "feature_results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Feature Results",
            "type": "array"
          },
          "latest_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Score"
          },
          "score_stream_count": {
            "title": "Score Stream Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "status",
          "score_stream_count",
          "evaluated_features"
        ],
        "title": "FeatureDriftResponse",
        "type": "object"
      },
      "FeedbackCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "agreed_with_model": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agreed With Model"
          },
          "decision": {
            "$ref": "#/components/schemas/OfficerDecision"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "prediction_id": {
            "title": "Prediction Id",
            "type": "string"
          }
        },
        "required": [
          "prediction_id",
          "decision"
        ],
        "title": "FeedbackCreateRequest",
        "type": "object"
      },
      "FeedbackListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FeedbackResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "FeedbackListResponse",
        "type": "object"
      },
      "FeedbackResponse": {
        "additionalProperties": false,
        "properties": {
          "agreed_with_model": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agreed With Model"
          },
          "amends_feedback_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amends Feedback Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/OfficerDecision"
          },
          "detector_state": {
            "$ref": "#/components/schemas/DriftStatusResponse"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_current": {
            "title": "Is Current",
            "type": "boolean"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "officer_id": {
            "title": "Officer Id",
            "type": "string"
          },
          "prediction_id": {
            "title": "Prediction Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "prediction_id",
          "officer_id",
          "version",
          "amends_feedback_id",
          "is_current",
          "decision",
          "agreed_with_model",
          "note",
          "detector_state",
          "created_at"
        ],
        "title": "FeedbackResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelMetadataResponse": {
        "additionalProperties": false,
        "properties": {
          "feature_count": {
            "title": "Feature Count",
            "type": "integer"
          },
          "feature_schema_sha256": {
            "title": "Feature Schema Sha256",
            "type": "string"
          },
          "feature_view": {
            "title": "Feature View",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "prediction_threshold": {
            "title": "Prediction Threshold",
            "type": "number"
          },
          "synthetic_demo": {
            "default": false,
            "title": "Synthetic Demo",
            "type": "boolean"
          }
        },
        "required": [
          "model_version",
          "feature_view",
          "prediction_threshold",
          "feature_count",
          "feature_schema_sha256"
        ],
        "title": "ModelMetadataResponse",
        "type": "object"
      },
      "MonitoringHistoryResponse": {
        "additionalProperties": false,
        "properties": {
          "snapshots": {
            "items": {
              "$ref": "#/components/schemas/MonitoringHistorySnapshotResponse"
            },
            "title": "Snapshots",
            "type": "array"
          }
        },
        "title": "MonitoringHistoryResponse",
        "type": "object"
      },
      "MonitoringHistorySnapshotResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "snapshot": {
            "$ref": "#/components/schemas/DriftStatusResponse"
          },
          "source": {
            "$ref": "#/components/schemas/MonitoringSnapshotSource"
          }
        },
        "required": [
          "id",
          "source",
          "snapshot",
          "created_at"
        ],
        "title": "MonitoringHistorySnapshotResponse",
        "type": "object"
      },
      "MonitoringSnapshotSource": {
        "enum": [
          "score_observation",
          "feature_drift"
        ],
        "title": "MonitoringSnapshotSource",
        "type": "string"
      },
      "OfficerDecision": {
        "enum": [
          "approve",
          "decline",
          "escalate"
        ],
        "title": "OfficerDecision",
        "type": "string"
      },
...
      "TrainingRunsResponse": {
        "additionalProperties": false,
        "properties": {
          "runs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Runs",
            "type": "array"
          },
          "synthetic_demo": {
            "default": false,
            "title": "Synthetic Demo",
            "type": "boolean"
          }
        },
        "title": "TrainingRunsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Human-in-the-loop loan default decision support with auditable model scores, explanations, officer feedback, and drift state. Decision support only; a qualified human makes the final decision.",
    "title": "Drift Aware Loan Default API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/feedback": {
      "get": {
        "operationId": "list_feedback_api_v1_feedback_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "prediction_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prediction Id"
            }
          },
          {
            "in": "query",
            "name": "decision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OfficerDecision"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Feedback",
        "tags": [
          "feedback"
        ]
      },
      "post": {
        "operationId": "create_feedback_api_v1_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Feedback",
        "tags": [
          "feedback"
        ]
      }
    },
    "/api/v1/model": {
      "get": {
        "operationId": "get_model_metadata_api_v1_model_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelMetadataResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Model Metadata",
        "tags": [
          "model"
        ]
      }
    },
    "/api/v1/monitoring/feature-drift": {
      "post": {
        "operationId": "feature_drift_api_v1_monitoring_feature_drift_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureDriftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureDriftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Feature Drift",
        "tags": [
          "monitoring"
        ]
      }
    },
    "/api/v1/monitoring/history": {
      "get": {
        "description": "Return the most recent persisted monitoring events in chronological order.",
        "operationId": "monitoring_history_api_v1_monitoring_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Monitoring History",
        "tags": [
          "monitoring"
        ]
      }
    },
    "/api/v1/monitoring/status": {
      "get": {
        "operationId": "monitoring_status_api_v1_monitoring_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriftStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Monitoring Status",
        "tags": [
          "monitoring"
        ]
      }
    },
...
    "/api/v1/training-runs": {
      "get": {
        "description": "Expose recorded run metadata; model training itself remains a CLI job.",
        "operationId": "get_training_runs_api_v1_training_runs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingRunsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Training Runs",
        "tags": [
          "model"
        ]
      }
    },
...
  }
}